patches: Fix condition in dahdicleanup.diff.
In dahdi_handle_event, the variable 'x' was, in the original iteration
of this patch, guarded by !defined(SIG_ANALOG_HANDLED), but on systems
with libpri, this would cause compilation failure:

chan_dahdi.c: In function 'dahdi_handle_event':
chan_dahdi.c:7966:68: error: 'x' undeclared (first use in this function)
 7966 | if (ioctl(p->subs[idx].dfd, DAHDI_DIALING, &x) == -1) {

This was thus not guarded by a define at all, which usually works,
except this can also cause compilation failure when libpri is not present:

chan_dahdi.c: In function 'dahdi_handle_event':
chan_dahdi.c:7840:18: error: unused variable 'x' [-Werror=unused-variable]
 7840 | int res, x;

The proper condition for when x needs to exist is when either SIG_ANALOG_HANDLED
is not defined (i.e. the original code, where duplicate analog handling code
would use x) or when HAVE_PRI is defined, since this code is not dead and also
uses x. The warning about unused variable would happen only when HAVE_PRI
was not defined, and thus with the dead code not being compiled, x would not
see any usage.

PHREAKSCRIPT-63 #close

+++ phreaknet.sh
@@ -2137,21 +2137,25 @@ install_dahdi() {
 	fi
 
 	# LibPRI # https://gist.github.com/debuggerboy/3028532
-	cd $AST_SOURCE_PARENT_DIR
-	wget http://downloads.asterisk.org/pub/telephony/libpri/releases/${LIBPRI_SOURCE_NAME}.tar.gz
-	tar -zxvf ${LIBPRI_SOURCE_NAME}.tar.gz
-	rm ${LIBPRI_SOURCE_NAME}.tar.gz
-	cd ${LIBPRI_SOURCE_NAME}
-	$AST_MAKE && $AST_MAKE install
+	if [ "$INSTALL_LIBPRI" != "0" ]; then
+		cd $AST_SOURCE_PARENT_DIR
+		wget http://downloads.asterisk.org/pub/telephony/libpri/releases/${LIBPRI_SOURCE_NAME}.tar.gz
+		tar -zxvf ${LIBPRI_SOURCE_NAME}.tar.gz
+		rm ${LIBPRI_SOURCE_NAME}.tar.gz
+		cd ${LIBPRI_SOURCE_NAME}
+		$AST_MAKE && $AST_MAKE install
+	fi
 
 	# LibSS7
-	cd $AST_SOURCE_PARENT_DIR
-	wget https://github.com/asterisk/libss7/archive/refs/tags/${LIBSS7_VERSION}.tar.gz
-	tar -zxvf ${LIBSS7_VERSION}.tar.gz
-	rm ${LIBSS7_VERSION}.tar.gz
-	ls -la
-	cd libss7-${LIBSS7_VERSION}
-	$AST_MAKE && $AST_MAKE install
+	if [ "$INSTALL_LIBSS7" != "0" ]; then
+		cd $AST_SOURCE_PARENT_DIR
+		wget https://github.com/asterisk/libss7/archive/refs/tags/${LIBSS7_VERSION}.tar.gz
+		tar -zxvf ${LIBSS7_VERSION}.tar.gz
+		rm ${LIBSS7_VERSION}.tar.gz
+		ls -la
+		cd libss7-${LIBSS7_VERSION}
+		$AST_MAKE && $AST_MAKE install
+	fi
 
 	# Wanpipe
 	if [ "$DAHDI_WANPIPE" = "1" ]; then
@@ -2465,7 +2469,7 @@ phreak_patches() { # $1 = $PATCH_DIR, $2 = $AST_SRC_DIR
 	asterisk_pr_unconditional 1089 # app_sms: Ignore false positive gcc warning
 	#asterisk_pr_unconditional 272 # Call Waiting Deluxe. This also now conflicts (with the latest revisions), so temp. disabled.
 	#asterisk_pr_unconditional 292 # GROUP VARs # Disabled temporarily as patch does not apply anymore
-	git_custom_patch "https://code.phreaknet.org/asterisk/dahdicleanup.diff"
+	git_patch "dahdicleanup.diff"
 
 	if [ $AST_MAJOR_VER -lt 21 ]; then
 		if [ "$EXTERNAL_CODECS" = "1" ]; then
@@ -3525,7 +3529,6 @@ elif [ "$cmd" = "install" ]; then
 		$AST_MAKE -j$(nproc) ASTCFLAGS="-w" utils # for astdb2sqlite3.c, #warning usage of non-standard #include <sys/cdefs.h> is deprecated [-Werror=cpp]
 		$AST_MAKE -j$(nproc) ASTCFLAGS="-w" addons # for format_mp3, error: #warning redirecting incorrect #include <sys/signal.h> to <signal.h> [-Werror=cpp]
 		$AST_MAKE -j$(nproc) ASTCFLAGS="-w" main # for ast_expr2.c, #warning usage of non-standard #include <sys/cdefs.h> is deprecated [-Werror=cpp]
-		$AST_MAKE -j$(nproc) ASTCFLAGS="-w" channels # XXX: Temporary, for chan_dahdi.c:7840:18: error: unused variable 'x' [-Werror=unused-variable]
 	fi
 
 	# Compile Asterisk. This is the longest step, if you are installing for the first time.
